apps/fetcher/requirements.txt
requests
beautifulsoup4
python-dotenv
upstash-redis